Created functions
Instead of just having this as a script, I created a number of functions that do the thing that we're interested in. I also removed the sequential aspect entirely, so that it does not gradually feed successes and failures into the fit. While that is an interesting exercise in itself, you can do that more pedagogically by showing a sequence of appropriately chosen plots from this thing.