Added frameless option, changed format to png
It is now possible to make a plot which only contains the function curve. This is intended to allow the user to plot a sequence of updates to P(p|n, k), without unnecessary clutter. I have also changed format from jpg to png, since there were visible compression artifacts.