Started work on flat_prior module
This module is intended to demonstrate a weird thing that can happen when fitting parameters, namely that different statisticians get different results despite using the same model and the same data. Within a Bayesian framework it is relatively easy to see why this happens, but within a frequentist framework it is harder - mostly because frequentists tend to believe their framework is immune to this sort of thing.